Update Modal js
Зробив один файл js для модалок, та оновив селектори визову модалок на html сторінках. Видалив зайві js скріпти.
