+++ src/hotel.html
@@ -25,6 +25,5 @@
     <!-- Base js -->
     <script type="module" src="main.js"></script>
     <script src="./js/menu-active.js"></script>
-    <script src="./js/modal-hotel.js"></script>
   </body>
 </html>
