+++ src/index.html
@@ -53,7 +53,6 @@
 
     <!-- Base js -->
     <script type="module" src="/main.js"></script>
-    <script src="./js/modal-index.js"></script>
     <script src="./js/menu-active-index.js"></script>
   </body>
 </html>
