+++ src/main.js
@@ -2,3 +2,4 @@ import './js/mobile-menu';
 import './js/main-restaurant-slick';
 import './js/room-hero-slider';
 import './js/slick';
+import './js/modal';
