+++ src/partials/hotel/double-luxury-room.html
@@ -107,7 +107,7 @@
       </div>
       <div class="hotel-luxery-btn-wrap">
         <button type="button" class="hotel-luxury-btn"></button>
-        <button type="button" class="hotel-luxury-btn"></button>
+        <button type="button" class="hotel-luxury-btn" data-price-open></button>
       </div>
    </div>
   </div>
