+++ src/partials/hotel/four-seater-room.html
@@ -112,7 +112,7 @@
       </div>
       <div class="four-seater-room-button">
         <a class="four-seater-room-btn" href="/src/room.html">More details</a>
-        <button class="four-seater-room-btn open-modal-price" type="button" data-price-open4>
+        <button class="four-seater-room-btn" type="button" data-price-open>
           View prices
         </button>
       </div>
