+++ src/partials/hotel/hero-hotel.html
@@ -3,8 +3,8 @@
     <div class="hotel-hero-wrap">
       <h1 class="hero-hotel-title">hotel</h1>
       <div class="hotel-hero-btn-wrap">
-        <button type="button" class="hotel-hero-btn">book a room</button>
-        <button type="button" class="hotel-hero-btn">view prices</button>
+        <button type="button" class="hotel-hero-btn" data-room-open>book a room</button>
+        <button type="button" class="hotel-hero-btn" data-price-open>view prices</button>
       </div>
     </div>
     <div class="hotel-hero-slider">
