+++ src/partials/hotel/triple-room.html
@@ -122,7 +122,7 @@
       </div>
       <div class="triple-room-button">
         <a class="triple-room-btn" href="/src/room.html">More details</a>
-        <button class="triple-room-btn" type="button" data-price-open3>View prices</button>
+        <button class="triple-room-btn" type="button" data-price-open>View prices</button>
       </div>
     </div>
   </div>
