+++ src/partials/main/restaurant-details.html
@@ -250,7 +250,7 @@
       </div>
     </div>
     <div class="button__rest_position">
-      <button type="button" class="button__rest_main" data-table-open2>
+      <button type="button" class="button__rest_main" data-table-open>
         book a table
       </button>
       <button type="button" class="button__rest_main">view the menu</button>
