+++ src/partials/restaurant/restaurant-menu.html
@@ -275,6 +275,6 @@
   </ul>
 </div>
 </div>
-<button type="button" class="menu-btn open-modal-table">book a table</button>
+<button type="button" class="menu-btn" data-table-open>book a table</button>
 </div>
 </section>
