+++ src/restaurant.html
@@ -23,6 +23,5 @@
     <!-- Base js -->
     <script type="module" src="main.js"></script>
     <script src="./js/menu-active.js"></script>
-    <script src="./js/modal-restaurant.js"></script>
   </body>
 </html>
