+++ src/room.html
@@ -50,6 +50,5 @@
     <!-- Base js -->
     <script type="module" src="main.js"></script>
     <script src="./js/menu-active.js"></script>
-    <script src="./js/modal-room.js"></script>
   </body>
 </html>
